Split pkg/server/server.go into logical files
Pure reorganization, no behavior changes. server.go grew to ~3.1k lines
with 89 functions spanning OAuth, JWE, client resolution, dynamic tools,
OpenAPI handlers, query parsing, and settings. Review and further work
(Auth0 OAuth flow) land on a much smaller surface now.

Layout:
- server.go (491 lines) — struct, constructor, Add/Register, HandleExecuteQuery,
  HandleReadOnlyQuery, context key constants
- server_auth_oauth.go (548) — OAuth validation, JWT parsing, JWKS/OpenID discovery
- server_auth_jwe.go (99) — JWE token extraction and claim parsing
- server_client.go (308) — ClickHouse client resolution from JWE/OAuth context
- server_dynamic_tools.go (910) — discovery and handlers for dynamic read/write tools
- server_openapi.go (415) — OpenAPI REST handlers
- server_query.go (150) — SELECT detection and AST-based blocked-clause checks
- server_resources.go (163) — MCP resources and prompts
- server_settings.go (169) — tool_input_settings allowlist and parsing

Verified with go build, go vet, and the full pkg/server test suite.
The only failing test (TestOAuthOpenAPIFullFlow/complete_oauth_openapi_flow)
fails identically on the pre-split branch due to a sandbox isolator
restriction on ExtraHosts — unrelated to the split.
